ReflexView Constructor.
Added a new code inside the constructor to handle user prefereces and link the class to the View elements. Overrided the OnSizeChanged method to read the total space in the screen. Added the addNewSpot method that creates a random icon in the view, adds it to the queue and inflates the laout.