Animations handler.
Added a new code to handle animate and its callbacks. Added a new method to handle what to do when the user misses a click. Added a logic to store the new highscore, if the user actual score is higher then the previous one. Added a method to cancel animations when the game hits its game over state.