Game controller methos.
Added the method pause, to remove any objects in use, including the soundpool, while the game is paused. Added the method resume, to return the soundpool and initialize the sounds again. Added the mehot resetGame. 

Added the on Touch event to check if the user missed the spot. Added a new control to make the animations go faster by each correct touch from the user.